[TAN-1482] Changed importer classes
