+++ back/engines/commercial/bulk_import_ideas/app/policies/bulk_import_ideas/patches/project_policy.rb
@@ -7,12 +7,8 @@ module BulkImportIdeas
         active_moderator?
       end
 
-      def example_xlsx?
+      def export_form?
         active_moderator?
-      end
-
-      def to_pdf?
-        active_admin?
       end
 
       def draft_ideas?
