+++ back/engines/commercial/bulk_import_ideas/config/routes.rb
@@ -4,11 +4,10 @@ BulkImportIdeas::Engine.routes.draw do
   namespace :web_api do
     namespace :v1 do
       resources :phases do
-        post 'importer/bulk_create/idea/xlsx', on: :member, to: 'import_ideas#bulk_create'
-        get 'importer/export_form/idea/xlsx', on: :member, to: 'import_ideas#example_xlsx'
-        get 'importer/export_form/idea/pdf', on: :member, to: 'import_ideas#to_pdf'
-        get 'importer/draft/idea', on: :member, to: 'import_ideas#draft_ideas'
-        patch 'importer/approve_all/idea', on: :member, to: 'import_ideas#approve_all'
+        post 'importer/bulk_create/:model/:format', on: :member, to: 'import_ideas#bulk_create'
+        get 'importer/export_form/:model/:format', on: :member, to: 'import_ideas#export_form' # pdf or xlsx
+        get 'importer/draft/:model', on: :member, to: 'import_ideas#draft_ideas'
+        patch 'importer/approve_all/:model', on: :member, to: 'import_ideas#approve_all'
         post 'importer/create_user', on: :member, to: 'phase_users#create_user'
       end
       resources :idea_imports, on: :member, to: 'import_ideas#show_idea_import'
