+++ back/engines/commercial/bulk_import_ideas/lib/tasks/bulk_import_ideas_tasks.rake
@@ -9,12 +9,12 @@ namespace :bulk_import do
   desc 'Imports ideas from a csv file, as specified by the path argument, into the tenant specified by the host.'
   task :ideas, %i[url host] => [:environment] do |_t, args|
     tenant = Tenant.find_by host: args[:host]
-    service = BulkImportIdeas::ImportIdeasService.new
+    service = BulkImportIdeas::IdeaImporter.new
 
     tenant.switch do
       xlsx = CSV.parse open(args[:url]).read, headers: true, col_sep: ',', converters: []
       idea_rows = service.xlsx_to_idea_rows xlsx
-      service.import_ideas idea_rows
+      service.import idea_rows
     end
   end
 end
