+++ back/engines/commercial/bulk_import_ideas/spec/services/idea_pdf_form_exporter_spec.rb
@@ -2,10 +2,10 @@
 
 require 'rails_helper'
 
-describe BulkImportIdeas::PrintCustomFieldsService do
+describe BulkImportIdeas::IdeaPdfFormExporter do
   let(:project) { create(:single_phase_ideation_project) }
   let(:custom_form) { create(:custom_form, participation_context: project) }
-  let(:service) { described_class.new project, custom_form.custom_fields, 'en', false }
+  let(:service) { described_class.new project, 'en', false }
 
   before do
     # Custom fields
@@ -30,9 +30,9 @@ describe BulkImportIdeas::PrintCustomFieldsService do
     create(:custom_field_option, custom_field: another_select_field, key: 'no', title_multiloc: { 'en' => 'No' })
   end
 
-  describe 'create_pdf' do
-    it 'prints a PDF form' do
-      expect(service.create_pdf).not_to be_nil
+  describe 'export' do
+    it 'returns a printable PDF form' do
+      expect(service.export).not_to be_nil
     end
   end
